requirements.txt
django==1.11
django-crispy-forms==1.7.2